devex: fix paths after files moved
